+++ NEWS.md
@@ -1,3 +1,10 @@
+# penalizedSVM 1.2.0
+
+* Update maintainer email
+* Add unit tests
+* Fix package to get rid of CRAN check notes
+* Add package doi
+
 # penalizedSVM 1.1.4
 
 * Fix package to get rid of CRAN check notes